Refactor to move functions to utils.js
Moving these functions just to make the index.js file simpler.
